+++ wgpu-hal/src/gles/adapter.rs
@@ -382,7 +382,7 @@ impl super::Adapter {
             // This is true, even though instance index will not be properly adjusted.
             //
             // This is because the OpenGL backend does not advertise support for non-zero `base_instance`.
-            | wgt::DownlevelFlags::VERTEX_AND_INSTANCE_INDEX_RESPECTS_RESPECTIVE_INDIRECT_BASE;
+            | wgt::DownlevelFlags::VERTEX_AND_INSTANCE_INDEX_RESPECTS_RESPECTIVE_INDIRECT_FIRST;
         downlevel_flags.set(wgt::DownlevelFlags::COMPUTE_SHADERS, supports_compute);
         downlevel_flags.set(
             wgt::DownlevelFlags::FRAGMENT_WRITABLE_STORAGE,
